A little bit of refactoring to match already existing in the project design pattern
